+++ app/src/main/res/layout/activity_home_screen.xml
@@ -65,8 +65,8 @@
 
     <Button
         android:id="@+id/SettingsBtn"
-        android:layout_width="53dp"
-        android:layout_height="48dp"
+        android:layout_width="51dp"
+        android:layout_height="58dp"
         android:layout_marginStart="-30dp"
         android:layout_marginTop="-10dp"
         android:layout_marginEnd="20dp"
