+++ app/src/main/res/layout/activity_profile_screen.xml
@@ -96,22 +96,5 @@
         app:layout_constraintTop_toBottomOf="@+id/edit_gift_ideas_opt" />
 
 
-    <Button
-        android:id="@+id/button_log_out"
-        android:layout_width="163dp"
-        android:layout_height="52dp"
-        android:layout_marginStart="124dp"
-        android:layout_marginTop="22dp"
-        android:layout_marginEnd="124dp"
-        android:fontFamily="sans-serif-black"
-        android:text="Log Out"
-        android:textAllCaps="false"
-        android:textSize="20dp"
-        app:backgroundTint="#4B4B4B"
-        app:cornerRadius="15dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/anonymity_opt" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
